+++ eds-aic-20230930/alibabacloud_eds_aic20230930/client.py
@@ -8075,6 +8075,8 @@ class Client(OpenApiClient):
             query['MaxSteps'] = request.max_steps
         if not DaraCore.is_null(request.schedule_id):
             query['ScheduleId'] = request.schedule_id
+        if not DaraCore.is_null(request.targets):
+            query['Targets'] = request.targets
         if not DaraCore.is_null(request.task_config_id):
             query['TaskConfigId'] = request.task_config_id
         if not DaraCore.is_null(request.timeout_seconds):
@@ -8115,6 +8117,8 @@ class Client(OpenApiClient):
             query['MaxSteps'] = request.max_steps
         if not DaraCore.is_null(request.schedule_id):
             query['ScheduleId'] = request.schedule_id
+        if not DaraCore.is_null(request.targets):
+            query['Targets'] = request.targets
         if not DaraCore.is_null(request.task_config_id):
             query['TaskConfigId'] = request.task_config_id
         if not DaraCore.is_null(request.timeout_seconds):