+++ eds-aic-20230930/alibabacloud_eds_aic20230930/models/_run_agent_task_response_body.py
@@ -85,6 +85,7 @@ class RunAgentTaskResponseBodyTasks(DaraModel):
         current_status: str = None,
         instance_id: str = None,
         running_at: str = None,
+        session_id: str = None,
         task_id: str = None,
         user_prompt: str = None,
     ):
@@ -100,6 +101,7 @@ class RunAgentTaskResponseBodyTasks(DaraModel):
         self.instance_id = instance_id
         # The time when the task was created, in ISO 8601 format.
         self.running_at = running_at
+        self.session_id = session_id
         # The task ID, which is globally unique.
         self.task_id = task_id
         # The user instruction in natural language. The Agent performs operations based on this instruction.
@@ -122,6 +124,9 @@ class RunAgentTaskResponseBodyTasks(DaraModel):
         if self.running_at is not None:
             result['RunningAt'] = self.running_at
 
+        if self.session_id is not None:
+            result['SessionId'] = self.session_id
+
         if self.task_id is not None:
             result['TaskId'] = self.task_id
 
@@ -140,6 +145,9 @@ class RunAgentTaskResponseBodyTasks(DaraModel):
 
         if m.get('RunningAt') is not None:
             self.running_at = m.get('RunningAt')
+
+        if m.get('SessionId') is not None:
+            self.session_id = m.get('SessionId')
 
         if m.get('TaskId') is not None:
             self.task_id = m.get('TaskId')